updating comment feature. I think it may work but my site is glitching rn
