+++ app.py
@@ -1,5 +1,5 @@
 from flask import Flask, render_template, request, redirect, flash, url_for, session
-from src.models import db, Post, Users
+from src.models import db, Post, Users, Comment
 from dotenv import load_dotenv
 import os
 from forms import RegistrationForm, LoginForm, SearchForm, EditProfileForm
@@ -239,24 +239,14 @@ def add_comment(post_id):
         flash('Comment cannot be empty', 'error')
         return redirect(url_for('get_single_post', post_id=post_id))
 
-    new_comment = Comment(comment=comment_text, user_id=current_user.id, post_id=post_id)
+    new_comment = Comment(comment=comment_text, id=current_user.id, post_id=post_id)
     db.session.add(new_comment)
     db.session.commit()
 
     flash('Comment added successfully', 'success')
     return redirect(url_for('get_single_post', post_id=post_id))
 
-class Comment(db.Model):
-    __tablename__ = 'comments'
 
-    comment_id = db.Column(db.Integer, primary_key=True)
-    comment = db.Column(db.String(255))
-
-    user_id = db.Column(db.Integer, db.ForeignKey('users.id'))
-    user = db.relationship('Users', backref='user_comments')
-
-    post_id = db.Column(db.Integer, db.ForeignKey('post.post_id'))
-    post = db.relationship('Post', backref='post_comments')
 
 
 @app.get('/post/<int:post_id>')
